Executing/src/requirement.txt
paho-mqtt == 1.5.1
numpy==1.13
pandas==0.20.2
scikit_learn== latest
